Add a scanner
This commit adds a basic scanner to the project. In particular, the `scan`
funciton in `scanner.h` dissects a string of Lox source code into a `vector`
of syntax `Token`s.

`scan` uses a `Scanner` helper class behind the scenes. Differing from the
canonical solution (using a DFA/NFA or generating code using e.g. `lex`),
`Scanner` parses the input string using two iterators and a switch statement.

`main` has been adapted to execute Lox code from a source file passed as a
command-line argument or directly from the shell in a REPL-like fashion.
"Executed" as of now means reading the source and displaying a list of scanned
tokens; this will change in future commits.