Groundwork for AST
This commit lays some foundational work for the Abstract Syntax Tree (AST),
which will be constructed by the parser. The AST is implemented by a number
of recursive structures defined in `ast.h`. Working with the tree is done via
the "Visitor" pattern. A basic `Printer` class defined in `ast-tools.h`
demonstrates this idea.

Additionally, the directory structure of the project has been refactored.
Header files shared among modules now reside in `include/`, while source code
and "private" headers are stored in `src/`. This may change in further commits.