Add parser for mathematical expressions
This commit adds a parser which transforms a vector of `Token`s into an
abstract syntax tree (AST). The parser is implemented in `parser.h`/`parser.cpp`
and exposes functionality through the `parse` function.

Similar to `scan`, `parse` uses a helper class to construct the AST. `Parser`
objects work by recursively calling member functions which correspond to
production rules in the grammar. If a syntax error is encountered, the parser
switches into panic mode and discards tokens until the start of a new
statement. For now, the parser can only work with mathematical expressions; this
will change in future commits.

Additionally, the overall project structure has been reworked; the code is now
grouped into modules residing in the `src` folder.