Add interpreter for mathematical expressions
This commit adds a basic interpreter which evaluates expressions given as an
AST. The interpreter is implemented in `Interpreter.h/`interpreter.cpp`.

The Interpreter class implements the Visitor interface of the Expression struct
hierarchy. The values of an expression is computed while traversing the
corresponding AST and printed to stdout afterwards. In future commits, this
will be extended to work with entire Lox statements.