Add functions
This commit adds support for functions. Functions can be declared by using
the `fun(<arguments>) <block>` syntax and are tied to the scope in which they
are defined in, thereby supporting closures. A native function `clock()` which
displays the system time since epoch in seconds has been added.

Due to several ownership issues that occured during this work, all objects are
now owned by shared_ptrs by default. A static member function `New()` has been
added to each data strucutre in order to facilitate construction of objects on
the heap. This may not be the most efficient solution, but it solves a series of
headaches – everything else is regarded "premaute optimization" for now.