Add resolver
This commit adds a dedicated name resolution step occuring between parsing and
interpretation. The resolver performs static analysis on the parsed AST and
instructs the interpreter in which environment a referenced variable is stored.
This solves an inconsistency bug that appears when a variable captured by an
anonymous function (i.e. a closure) is modified after the closure is defined.

To do so, the Visitor interfaces of the `Expression` and `Statement` strucutres
had to be modified; visitors now visit mutable References instead of const
references in order to take addresses of `Variable` expressions. The interpreter
uses these addresses as map keys to associate a variable with its containing
scope. A better solution would have been to provide hash functions for these
objects; this might be added in future commits.